Current state of syncChangesToServer and syncCollectionContents.
Also this renames getServerData() to getServerRow() for consistency.

+++ src/com/morphoss/acal/service/SyncCollectionContents.java
@@ -57,7 +57,7 @@ public class SyncCollectionContents extends ServiceJob {
 	 * @param collectionId
 	 * @param forceSync
 	 */
-	public SyncCollectionContents(int collectionId, boolean forceSync ) {
+	public SyncCollectionContents(long collectionId, boolean forceSync ) {
 		this.collectionId = collectionId;
 		request = new RRSyncCollectionContents(collectionId, forceSync);
 		this.TIME_TO_EXECUTE = 0;
